chore: Update JS packaging
